+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Breaking changes
   - Removed `MagicAuth.delete_all_sessions_by_token/1`. Use `MagicAuth.log_out/1` instead.
   - Removed `MagicAuth.delete_all_sessions_by_email/1`. Use `MagicAuth.log_out_all/1` instead.
+  - Magic Auth now redirects to the configured log in page instead of `/` after log out.
 
 ### How use automatic user loading feature on legacy project